Move BaseCMIP6Model and BaseCMIP6Record into plugins/cmip6/base; update imports; tests+lint green
